requirements.txt
flask==2.3.3
spotipy==2.23.0
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
selenium==4.15.0
webdriver-manager==4.0.1
gunicorn==21.2.0
torch>=1.11.0
numpy>=1.21.0
lxml>=4.6.0
urllib3>=1.26.0
charset-normalizer>=2.0.0
openai>=1.0.0
